@@ -2,6 +2,7 @@
   "name": "portfolio",
   "version": "0.0.1",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite dev",
     "build": "vite build",
@@ -12,35 +13,39 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-vercel": "^5.1.0",
-    "@sveltejs/enhanced-img": "^0.1.8",
-    "@sveltejs/kit": ">=2.4.3",
-    "@sveltejs/vite-plugin-svelte": "^3.0.0",
-    "@types/node": "^20.3.3",
+    "@eslint/compat": "^1.2.3",
+    "@sveltejs/adapter-vercel": "^5.5.0",
+    "@sveltejs/enhanced-img": "^0.4.4",
+    "@sveltejs/kit": "^2.5.27",
+    "@sveltejs/vite-plugin-svelte": "^5.0.0",
+    "@types/node": "^22.0.0",
     "@typescript-eslint/eslint-plugin": "^6.10.0",
     "@typescript-eslint/parser": "^6.10.0",
-    "autoprefixer": "^10.4.17",
-    "eslint": "^8.53.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-svelte": "^2.35.0",
-    "mdsvex": "^0.11.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^9.17.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-svelte": "^2.45.1",
+    "globals": "^15.0.0",
+    "mdsvex": "^0.11.2",
     "postcss": "^8.4.35",
-    "prettier": "^3.2.5",
-    "prettier-plugin-svelte": "^3.2.1",
-    "prettier-plugin-tailwindcss": "^0.5.11",
-    "svelte": "^4.2.3",
-    "svelte-check": "^3.0.1",
-    "tailwindcss": "^3.4.1",
+    "prettier": "^3.3.2",
+    "prettier-plugin-svelte": "^3.2.6",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "svelte": "^5.0.0",
+    "svelte-check": "^4.0.0",
+    "tailwindcss": "^3.4.9",
     "tslib": "^2.4.1",
-    "typescript": "^5.2.2",
-    "vite": "^5.0.0"
+    "typescript": "^5.5.0",
+    "typescript-eslint": "^8.0.0",
+    "vite": "^6.0.0"
   },
-  "type": "module",
   "dependencies": {
+    "@tailwindcss/typography": "^0.5.15",
     "@vercel/analytics": "^0.1.11",
     "@vercel/speed-insights": "^1.0.3",
-    "pixi.js": "^8.1.0",
-    "svelte-i18n": "^4.0.0"
-  },
-  "packageManager": "pnpm@8.15.7+sha256.50783dd0fa303852de2dd1557cd4b9f07cb5b018154a6e76d0f40635d6cee019"
+    "lucide-svelte": "^0.469.0",
+    "postcss-easing-gradients": "^3.0.1",
+    "svelte-i18n": "^4.0.0",
+    "vite-plugin-toml": "^0.7.0"
+  }
 }
