@@ -1,6 +1,10 @@
 <script lang="ts">
   import type { PageServerData } from "./$types";
-  export let data: PageServerData;
+  interface Props {
+    data: PageServerData;
+  }
+
+  let { data }: Props = $props();
 </script>
 
 <svelte:head>
