@@ -1,14 +1,11 @@
 import { init, addMessages, locale } from "svelte-i18n";
 import { derived } from "svelte/store";
 
-import en from "../locales/en.json";
-import es from "../locales/es.json";
+import en from "../locales/en.toml";
+import es from "../locales/es.toml";
 
 addMessages("en", en);
 addMessages("es", es);
-
-//console.log(es);
-//console.log(en);
 
 const defaultLocale = "en";
 
